# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GO API using Gin framework.",
        "title": "GO API Course",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/change-password": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ChangePassword",
                "parameters": [
                    {
                        "description": "Change Password payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.ChangePasswordPayload"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "User payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserPayload"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/signup": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Signup",
                "parameters": [
                    {
                        "description": "User payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserPayload"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/validate": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Validate",
                "responses": {}
            }
        }
    },
    "definitions": {
        "requests.ChangePasswordPayload": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.doe@mail.com"
                },
                "new_password": {
                    "type": "string",
                    "example": "password2"
                },
                "old_password": {
                    "type": "string",
                    "example": "password1"
                }
            }
        },
        "requests.UserPayload": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.doe@mail.com"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                }
            }
        }
    }
}